Avoid split packages in samples
In a multi project, all subprojects used the same package,
which is a bad practice.

This commit also removes some redundant headers and imports.
