+++ subprojects/docs/src/samples/groovy/application/groovy/build.gradle
@@ -15,5 +15,5 @@ dependencies {
 }
 
 application {
-    mainClassName = 'org.gradle.sample.Main'
+    mainClassName = 'org.gradle.sample.app.Main'
 }
