+++ subprojects/docs/src/samples/groovy/application/kotlin/build.gradle.kts
@@ -15,5 +15,5 @@ dependencies {
 }
 
 application {
-    mainClassName = "org.gradle.sample.Main"
+    mainClassName = "org.gradle.sample.app.Main"
 }
