+++ subprojects/docs/src/samples/java/application/groovy/build.gradle
@@ -6,5 +6,5 @@ version = '1.0.2'
 group = 'org.gradle.sample'
 
 application {
-    mainClassName = 'org.gradle.sample.Main'
+    mainClassName = 'org.gradle.sample.app.Main'
 }
