+++ subprojects/docs/src/samples/java/components-with-junit-5-tests/groovy/application/build.gradle
@@ -10,7 +10,7 @@ dependencies {
 }
 
 application {
-    mainClassName = 'org.gradle.sample.Main'
+    mainClassName = 'org.gradle.sample.app.Main'
 }
 
 test {
