+++ subprojects/docs/src/samples/java/components-with-junit-5-tests/kotlin/application/build.gradle.kts
@@ -10,7 +10,7 @@ dependencies {
 }
 
 application {
-    mainClassName = "org.gradle.sample.Main"
+    mainClassName = "org.gradle.sample.app.Main"
 }
 
 tasks.named<Test>("test") {
