+++ subprojects/docs/src/samples/jvm-components-with-additional-test-types/groovy/application/build.gradle
@@ -10,7 +10,7 @@ dependencies {
 }
 
 application {
-    mainClassName = 'org.gradle.sample.Main'
+    mainClassName = 'org.gradle.sample.app.Main'
 }
 
 sourceSets {
