+++ subprojects/docs/src/samples/jvm-components-with-additional-test-types/kotlin/application/build.gradle.kts
@@ -10,7 +10,7 @@ dependencies {
 }
 
 application {
-    mainClassName = "org.gradle.sample.Main"
+    mainClassName = "org.gradle.sample.app.Main"
 }
 
 sourceSets {
