+++ subprojects/docs/src/samples/jvm-multi-project-build/kotlin/application/build.gradle.kts
@@ -7,5 +7,5 @@ dependencies {
 }
 
 application {
-    mainClassName = "org.gradle.sample.Main"
+    mainClassName = "org.gradle.sample.app.Main"
 }
