+++ subprojects/docs/src/samples/kotlin/application/kotlin/build.gradle.kts
@@ -7,7 +7,7 @@ version = "1.0.2"
 group = "org.gradle.sample"
 
 application {
-    mainClassName = "org.gradle.sample.MainKt"
+    mainClassName = "org.gradle.sample.app.MainKt"
 }
 
 repositories {
