--- subprojects/docs/src/samples/templates/groovy-application/src/main/groovy/org/gradle/sample/Main.groovy
@@ -1,15 +1,0 @@
-package org.gradle.sample
-
-import org.gradle.sample.LinkedList
-
-import static org.gradle.sample.StringUtils.join
-import static org.gradle.sample.StringUtils.split
-import static org.gradle.sample.MessageUtils.getMessage
-
-class Main {
-    static void main(String[] args) {
-        LinkedList tokens
-        tokens = split(getMessage())
-        println(join(tokens))
-    }
-}
