+++ subprojects/docs/src/samples/templates/groovy-application/src/main/groovy/org/gradle/sample/app/Main.groovy
@@ -1,0 +1,13 @@
+package org.gradle.sample.app
+
+import static org.gradle.sample.utilities.StringUtils.join
+import static org.gradle.sample.utilities.StringUtils.split
+import static org.gradle.sample.app.MessageUtils.getMessage
+
+class Main {
+    static void main(String[] args) {
+        LinkedList tokens
+        tokens = split(getMessage())
+        println(join(tokens))
+    }
+}
