+++ subprojects/docs/src/samples/templates/groovy-application/src/main/groovy/org/gradle/sample/app/MessageUtils.groovy
@@ -1,4 +1,4 @@
-package org.gradle.sample
+package org.gradle.sample.app
 
 import groovy.transform.PackageScope
