+++ subprojects/docs/src/samples/templates/groovy-spock-test-for-application/src/test/groovy/org/gradle/sample/app/MessageUtilsTest.groovy
@@ -1,4 +1,4 @@
-package org.gradle.sample
+package org.gradle.sample.app
 
 import spock.lang.Specification
