+++ subprojects/docs/src/samples/templates/groovy-spock-test-for-list-library/src/test/groovy/org/gradle/sample/list/LinkedListTest.groovy
@@ -1,4 +1,4 @@
-package org.gradle.sample
+package org.gradle.sample.list
 
 import spock.lang.Specification
