+++ subprojects/docs/src/samples/templates/groovy-utilities-library/src/main/groovy/org/gradle/sample/utilities/JoinUtils.groovy
@@ -1,6 +1,7 @@
-package org.gradle.sample
+package org.gradle.sample.utilities
 
 import groovy.transform.PackageScope
+import org.gradle.sample.list.LinkedList
 
 @PackageScope
 class JoinUtils {
