+++ subprojects/docs/src/samples/templates/groovy-utilities-library/src/main/groovy/org/gradle/sample/utilities/SplitUtils.groovy
@@ -1,23 +1,7 @@
-/*
- * Copyright 2019 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.sample
+package org.gradle.sample.utilities
 
 import groovy.transform.PackageScope
-import org.gradle.sample.LinkedList
+import org.gradle.sample.list.LinkedList
 
 @PackageScope
 class SplitUtils {
