+++ subprojects/docs/src/samples/templates/groovy-utilities-library/src/main/groovy/org/gradle/sample/utilities/StringUtils.groovy
@@ -1,0 +1,13 @@
+package org.gradle.sample.utilities
+
+import org.gradle.sample.list.LinkedList
+
+class StringUtils {
+    static String join(LinkedList source) {
+        return JoinUtils.join(source)
+    }
+
+    static LinkedList split(String source) {
+        return SplitUtils.split(source)
+    }
+}
