+++ subprojects/docs/src/samples/templates/java-application/src/main/java/org/gradle/sample/app/Main.java
@@ -1,0 +1,15 @@
+package org.gradle.sample.app;
+
+import org.gradle.sample.list.LinkedList;
+
+import static org.gradle.sample.utilities.StringUtils.join;
+import static org.gradle.sample.utilities.StringUtils.split;
+import static org.gradle.sample.app.MessageUtils.getMessage;
+
+public class Main {
+    public static void main(String[] args) {
+        LinkedList tokens;
+        tokens = split(getMessage());
+        System.out.println(join(tokens));
+    }
+}
