+++ subprojects/docs/src/samples/templates/java-application/src/main/java/org/gradle/sample/app/MessageUtils.java
@@ -1,4 +1,4 @@
-package org.gradle.sample;
+package org.gradle.sample.app;
 
 class MessageUtils {
     public static String getMessage() {
