+++ subprojects/docs/src/samples/templates/java-junit5-integration-test-for-utilities-library/src/integrationTest/java/org/gradle/sample/integtest/utilities/StringUtilsIntegrationTest.java
@@ -1,0 +1,19 @@
+package org.gradle.sample.integtest.utilities;
+
+import org.junit.jupiter.api.Test;
+
+import org.gradle.sample.list.LinkedList;
+import org.gradle.sample.utilities.StringUtils;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class StringUtilsIntegrationTest {
+    @Test public void testSplit() {
+        LinkedList list = StringUtils.split("The dog is green");
+        assertEquals(4, list.size());
+        assertEquals("The", list.get(0));
+        assertEquals("dog", list.get(1));
+        assertEquals("is", list.get(2));
+        assertEquals("green", list.get(3));
+    }
+}
