+++ subprojects/docs/src/samples/templates/java-junit5-test-for-application/src/test/java/org/gradle/sample/app/MessageUtilsTest.java
@@ -1,0 +1,11 @@
+package org.gradle.sample.app;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class MessageUtilsTest {
+    @Test public void testGetMessage() {
+        assertEquals("Hello,      World!", MessageUtils.getMessage());
+    }
+}
