+++ subprojects/docs/src/samples/templates/java-junit5-test-for-list-library/src/test/java/org/gradle/sample/list/LinkedListTest.java
@@ -1,4 +1,4 @@
-package org.gradle.sample;
+package org.gradle.sample.list;
 
 import org.junit.jupiter.api.Test;
