+++ subprojects/docs/src/samples/templates/java-utilities-library/src/main/java/org/gradle/sample/utilities/SplitUtils.java
@@ -1,6 +1,6 @@
-package org.gradle.sample;
+package org.gradle.sample.utilities;
 
-import org.gradle.sample.LinkedList;
+import org.gradle.sample.list.LinkedList;
 
 class SplitUtils {
     public static LinkedList split(String source) {
