+++ subprojects/docs/src/samples/templates/kotlin-application/src/main/kotlin/org/gradle/sample/app/Main.kt
@@ -1,4 +1,6 @@
-package org.gradle.sample
+package org.gradle.sample.app
+
+import org.gradle.sample.utilities.StringUtils
 
 fun main() {
     val tokens = StringUtils.split(MessageUtils.getMessage())
