+++ subprojects/docs/src/samples/templates/kotlin-application/src/main/kotlin/org/gradle/sample/app/MessageUtils.kt
@@ -1,4 +1,4 @@
-package org.gradle.sample
+package org.gradle.sample.app
 
 class MessageUtils {
     companion object {
