+++ subprojects/docs/src/samples/templates/kotlin-list-library/src/main/kotlin/org/gradle/sample/list/LinkedList.kt
@@ -1,4 +1,4 @@
-package org.gradle.sample
+package org.gradle.sample.list
 
 class LinkedList {
     private var head: Node? = null
