+++ subprojects/docs/src/samples/templates/kotlin-utilities-library/src/main/kotlin/org/gradle/sample/utilities/SplitUtils.kt
@@ -1,4 +1,6 @@
-package org.gradle.sample
+package org.gradle.sample.utilities
+
+import org.gradle.sample.list.LinkedList
 
 class SplitUtils {
     companion object {
